// apps/pos/src/app/views/SelfServe/types.ts
export type PricingTypes = 'monthly' | 'lifetime' | 'free';

export type Gallery = {
  thumbnail: string;
  main: string;
  mobile: string;
};

export type PricingBreakupkeys = 'monthly' | 'lifetime' | 'setup_fee';

export type PricingBreakup = {
  key: PricingBreakupkeys;
  description: string;
  value: number;
  suffix: string;
  isExtraFee: boolean;
  isChargeableAtCheckout: boolean;
  prevValue: number | null;
  nextValue: number | null;
};

export type ProductDescriptionPricing = {
  name: string;
  type: PricingTypes;
  subText: string;
  breakups: PricingBreakup[];
};

export type FeatureGallery = {
  image: string;
  title: string;
  description: string;
  isImageFirst: boolean;
};

export type InfoBanner = {
  image: string;
  mobileImage: string;
  features: {
    icon: string;
    text: string;
  }[];
};

export type TechnicalSpecification = {
  category: string;
  value: string;
};

export type LinkedItem = {
  title: string;
  offerLabel: string;
  quantity: number;
  image: any;
};

export type ProductDescription = {
  gallery: Gallery[];
  code: string;
  name: string;
  productTitle: string;
  description: string;
  cartImage: string;
  pricing: ProductDescriptionPricing[];
  isPartnerPricing: boolean;
  maxOrder: number | null;
  featureGallery: FeatureGallery[];
  infoBanner: InfoBanner;
  technicalSpecifications: TechnicalSpecification[];
  offer: {
    offerText: string;
    pdpOfferText: string;
    partnerOfferText: string;
    partnerPdpOfferText: string;
  } | null;
  shouldShowProductVarietyTable?: boolean;
  linkedItems?: LinkedItem[];
  maxQuantityErrMsg?: (quantity: number | null) => string;
  rentalDiscountPeriod: number;
};